+++ semcore/skeleton/CHANGELOG.md
@@ -2,11 +2,11 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
-## [5.31.0] - 2024-06-28
+## [5.31.0] - 2024-07-12
 
 ### Added
 
-- `CigarettePackChartSkeleton` component.
+- `CompactHorizontalBarChartSkeleton` component.
 
 ## [5.30.0] - 2024-07-12